Fixing upload scripts

+++ media_doc.php
@@ -93,7 +93,7 @@ header("location: Login.php");
                     echo "<td align='center'>".$row['SIZE']."</td>";
 					echo "<td align='center'>".$row['SYNOPSIS']."</td>";
                     echo "<td align='center'>
-                <a href='delete_jae.php?num=".$row['CONTENTID']."'>DEL</a></td>";
+                <a href='delete_jae.php?num=".$row['CONTENTID']."' onclick="return confirm('Are you sure?')">DEL</a></td>";
 					echo "<td align='center'>
 					<a href='". $target_dir.$row['CONTENTTITLE']."'>View</a></td>";
